+++ packages/primitives/src/components/Dropdown/DropdownItem.tsx
@@ -1,0 +1,8 @@
+import { DropdownItemProps } from './dropdown.types';
+
+const DropdownItem = (props: DropdownItemProps) => {
+  const { children, ...restProps } = props;
+  return <li {...restProps}>{children}</li>;
+};
+
+export default DropdownItem;
