+++ packages/primitives/src/components/Dropdown/dropdown.types.ts
@@ -1,0 +1,13 @@
+import { MouseEvent, PropsWithChildren, ReactNode } from 'react';
+
+export type DropdownItemProps = PropsWithChildren & {
+  onClick?: (event?: MouseEvent<HTMLLIElement>) => void;
+};
+
+export type DropdownListProps = PropsWithChildren & {
+  transformOrigin: 'left' | 'right';
+};
+
+export type DropdownTriggerProps = {
+  trigger: ReactNode;
+};
